requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
alembic>=1.11.0
httpx>=0.24.0
apscheduler>=3.10.0
pydantic-settings>=2.0.0
pydantic>=2.0.0
jinja2>=3.1.0
pyyaml>=6.0
